+++ GraphQL/Basket/BasketItemDeleteInput.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace WeDoTakeawayAPI.GraphQL.Baskets
+namespace WeDoTakeawayAPI.GraphQL.Basket
 {
     public record BasketItemDeleteInput (
         Guid BasketId,
