+++ GraphQL/Basket/BasketItemInput.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace WeDoTakeawayAPI.GraphQL.Baskets
+namespace WeDoTakeawayAPI.GraphQL.Basket
 {
     public record BasketItemInput(
         Guid BasketId,
